pyproject.toml
 [project]
 name = "hotel-shopping-agent"
 version = "0.1.0"
 description = "Hotel Shopping AI Agent MVP (LangGraph + FastAPI + Postgres)"
 requires-python = ">=3.11"
 readme = "README.md"
 dependencies = [
   "fastapi>=0.110",
   "uvicorn[standard]>=0.27",
   "pydantic>=2.6",
   "pydantic-settings>=2.2",
   "httpx>=0.27",
   "langgraph>=0.2.0",
   "langchain-core>=0.3.0",
   "langchain-openai>=0.3.0",
   "openai>=1.40.0",
   "sqlalchemy[asyncio]>=2.0.25",
   "asyncpg>=0.29",
   "psycopg[binary]>=3.1",
   "alembic>=1.13",
   "python-dateutil>=2.9",
   "orjson>=3.9",
   "structlog>=24.1",
   "prometheus-client>=0.20",
   "opentelemetry-api>=1.24",
   "opentelemetry-sdk>=1.24",
   "opentelemetry-exporter-otlp>=1.24",
   "opentelemetry-instrumentation-fastapi>=0.45b0",
   "opentelemetry-instrumentation-httpx>=0.45b0",
   "opentelemetry-instrumentation-sqlalchemy>=0.45b0",
 ]

 [project.optional-dependencies]
 dev = [
   "pytest>=8.0",
   "pytest-asyncio>=0.23",
   "testcontainers>=4.7",
   "requests>=2.31",
   "ruff>=0.6",
 ]

 [tool.pytest.ini_options]
 testpaths = ["tests"]
 asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"

 [tool.ruff]
 line-length = 100
 target-version = "py311"
 extend-exclude = ["db/migrations/alembic/versions"]

 [build-system]
 requires = ["setuptools>=68"]
 build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["services*", "db*", "evals*"]

